scripts/preconditions/requirements.txt
google-auth~=1.6
